moved setup_g to source_app
